修复: 负数整数解析错误，新增连接池与复合查询支持，完善测试用例
1. 修复ColToString中整数类型转换未显式转换类型的问题
2. 新增查询连接池，支持并发查询控制
3. 修复分页查询、上下文查询的会话资源管理
4. 新增复合JOIN条件合并逻辑，修复视图连接查询问题
5. 新增并完善多项测试用例，包括负数整数、复合连接、并发查询等

+++ purego/props.go
@@ -200,9 +200,9 @@ func (mdb *MdbHandle) ColToString(buf []byte, start int, dataType int, size int)
 	case MDBByte:
 		return fmt.Sprintf("%d", buf[start])
 	case MDBInt:
-		return fmt.Sprintf("%d", GetInt16(buf, start))
+		return fmt.Sprintf("%d", int16(GetInt16(buf, start)))
 	case MDBLongInt:
-		return fmt.Sprintf("%d", GetInt32(buf, start))
+		return fmt.Sprintf("%d", int32(GetInt32(buf, start)))
 	case MDBFloat:
 		return fmt.Sprintf("%g", GetSingle(buf, start))
 	case MDBDouble:
